fix request changes
