+++ index.html
@@ -55,6 +55,6 @@
     </footer>
 
     
-    <script type="module" src="./modules/main.js"></script>
+    <script type="module" src="./index.js"></script>
 </body>
 </html>
