+++ modules/bookclass.js
@@ -1,17 +1,7 @@
 /* eslint max-classes-per-file: ["error", 2] */
 
 const bookContainer = document.querySelector('.books');
-export const renderElements = (arr, container) => {
-  container.innerHTML = '';
-  arr.forEach((element, index) => {
-    container.innerHTML += `
-          <div class="book-element">
-            <h2>${element.title}</h2>
-            <p>${element.author}</p>
-            <button type="button" data-id=${index} class="btn-rm"> Remove </button>
-          </div>`;
-  });
-};
+import {renderElements} from './element.js'
 export default class BookList {
   static getBooks = () => {
     let bookList = [];
